requirements.txt
fastapi
uvicorn
icecream
aioconsole
motor
pydantic 
passlib[bcrypt]
PyJWT
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv